started implementing a wrapper for polygon.io API endpoints on the Doggie Class.
I know there exists a polygon API python module but it produces JSON. I want it to produce pandas dataframes to be ready for action

I've also realized that Alpaca does not produce adjusted values and is therefore relatively useless for any long term analysis

If a company were to have a stock split down the middle, the price would change yet the value wouldn't. Alpaca is not sensitive to that

That's why the random portfolio function sometimes returns ridiculous values for expected return and therefore for Sharpe Ratio

I've found Sharpe Ratios of 80 in there... absolutely savage.